+++ src/features/blog/blogList/components/Item/Content.tsx
@@ -1,34 +1,30 @@
 import { Link } from 'react-router-dom';
 
 import Button from '@components/UI/Button';
-import CategoryList from '@features/blog/CategoryList';
-import MetaList from '@features/blog/MetaList';
-import type { Post } from '@typings/posts/post';
-import type { Category } from '@typings/posts/category';
-import type { Author } from '@typings/posts/author';
 import { textTruncateByWords } from '@utils/formatters';
 
+import CategoryList from '../../../CategoryList';
+import MetaList from '../../../MetaList';
+
+import type { Posts } from '../../schemas/posts.schema';
+
 interface ContentProps {
-  post: Post;
-  categoryMap: Record<number, Category>;
-  authorMap: Record<number, Author>;
+  post: Posts;
   excerptLength: number;
   contentClasses: string;
 }
 
-export default function Content({ post, categoryMap, authorMap, excerptLength, contentClasses }: ContentProps) {
-  const author = authorMap[post.authorId];
-
+export default function Content({ post, excerptLength, contentClasses }: ContentProps) {
   return (
     <div className={`flex flex-col ${contentClasses} gap-y-8 bg-white`}>
       <header className="flex flex-col gap-y-5">
-        {post.categories && post.categories.length > 0 && <CategoryList categories={post.categories} categoryMap={categoryMap} />}
+        {post.categories && <CategoryList categories={post.categories} />}
         <h2>
           <Link to={`/blog/${post.slug}`} className="transition-colors duration-300 ease-in-out hover:text-boulder focus:text-boulder">
             {post.title}
           </Link>
         </h2>
-        <MetaList author={author} date={post.date} />
+        <MetaList author={post.author} date={post.created_at} />
       </header>
       {post.excerpt && <p>{textTruncateByWords(post.excerpt, excerptLength)}</p>}
       <Button to={`/blog/${post.slug}`} className="max-w-fit">