+++ src/features/blog/blogList/components/pagination/Numbers.tsx
@@ -1,19 +1,19 @@
-interface NumbersProps {
+interface BlogListPaginationNumbersProps {
   totalPages: number;
-  currentPage: number;
-  handlePageChange: (pageNumber: number) => void;
+  blogListPage: number;
+  handleBlogListPageChange: (pageNumber: number) => void;
 }
 
-export default function Numbers({ totalPages, currentPage, handlePageChange }: NumbersProps) {
+export default function BlogListPaginationNumbers({ totalPages, blogListPage, handleBlogListPageChange }: BlogListPaginationNumbersProps) {
   return (
     <>
       {[...Array(totalPages)].map((_, i) => {
         const page = i + 1;
-        const numberClasses = currentPage === page ? 'border-shark text-shark' : 'border-transparent';
+        const numberClasses = blogListPage === page ? 'border-shark text-shark' : 'border-transparent';
         return (
           <li key={page}>
             <button
-              onClick={() => handlePageChange(page)}
+              onClick={() => handleBlogListPageChange(page)}
               className={`w-8 cursor-pointer border-b-1 p-2 text-center text-boulder transition-colors duration-300 ease-in-out hover:text-shark focus:text-shark ${numberClasses}`}
               aria-label={`Page ${page}`}
             >