+++ src/features/blog/blogList/components/pagination/Pagination.tsx
@@ -1,20 +1,25 @@
-import Previous from './Previous';
-import Next from './Next';
-import Numbers from './Numbers';
+import BlogListPaginationPrevious from './Previous';
+import BlogListPaginationNumbers from './Numbers';
+import BlogListPaginationNext from './Next';
 
-interface PaginationProps {
-  totalPages: number;
-  currentPage: number;
-  handlePageChange: (pageNumber: number) => void;
+interface BlogListPaginationProps {
+  postsCount?: number | null;
+  postsPerPage: number;
+  blogListPage: number;
+  handleBlogListPageChange: (pageNumber: number) => void;
 }
 
-export default function Pagination({ totalPages, currentPage, handlePageChange }: PaginationProps) {
+export default function BlogListPagination({ postsCount, postsPerPage, blogListPage, handleBlogListPageChange }: BlogListPaginationProps) {
+  if (!postsCount) return null;
+
+  const totalPages = Math.ceil(postsCount / postsPerPage);
+
   return (
     <nav aria-label="Post pagination">
       <ul className="mt-10 flex justify-center gap-x-4">
-        {currentPage !== 1 && <Previous currentPage={currentPage} handlePageChange={handlePageChange} />}
-        <Numbers totalPages={totalPages} currentPage={currentPage} handlePageChange={handlePageChange} />
-        {currentPage !== totalPages && <Next currentPage={currentPage} handlePageChange={handlePageChange} />}
+        {blogListPage !== 1 && <BlogListPaginationPrevious blogListPage={blogListPage} handleBlogListPageChange={handleBlogListPageChange} />}
+        <BlogListPaginationNumbers totalPages={totalPages} blogListPage={blogListPage} handleBlogListPageChange={handleBlogListPageChange} />
+        {blogListPage !== totalPages && <BlogListPaginationNext blogListPage={blogListPage} handleBlogListPageChange={handleBlogListPageChange} />}
       </ul>
     </nav>
   );